+++ src/components/data-management/ProjectMenu.jsx
@@ -1,9 +1,7 @@
 import React, {
-  useState, useEffect, useCallback, useMemo,
+  useState, useEffect, useCallback,
 } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import sha1 from 'crypto-js/sha1';
-import Hex from 'crypto-js/enc-hex';
 import {
   Space, Button, Tooltip, Popconfirm,
 } from 'antd';
@@ -16,6 +14,7 @@ import FileUploadModal from './FileUploadModal';
 import AnalysisModal from './AnalysisModal';
 import { processUpload } from '../../utils/upload/processUpload';
 import integrationTestIds from '../../utils/integrationTestIds';
+import generateGem2sHashParams from '../../utils/data-management/generateGem2sHashParams';
 
 const ProjectMenu = () => {
   const dispatch = useDispatch();
@@ -79,37 +78,6 @@ const ProjectMenu = () => {
 
     if (!experiments[experimentId]?.sampleIds.length > 0) return;
 
-    const generateGem2sHashParams = (project) => {
-      const experiment = experiments[experimentId];
-      const experimentSamples = project.samples.map((sampleUuid) => samples[sampleUuid]);
-
-      const samplesEntries = Object.entries(experimentSamples);
-
-      // Different sample order should not change the hash.
-      const orderInvariantSampleIds = [...experiment.sampleIds].sort();
-
-      const hashParams = {
-        organism: experiment.meta.organism,
-        input: { type: experiment.meta.type },
-        sampleIds: orderInvariantSampleIds,
-        sampleNames: orderInvariantSampleIds.map((sampleId) => samples[sampleId].name),
-      };
-
-      if (project.metadataKeys.length) {
-        hashParams.metadata = project.metadataKeys.reduce((acc, key) => {
-        // Make sure the key does not contain '-' as it will cause failure in GEM2S
-          const sanitizedKey = key.replace(/-+/g, '_');
-
-          acc[sanitizedKey] = samplesEntries.map(
-            ([, sample]) => sample.metadata[key] || DEFAULT_NA,
-          );
-          return acc;
-        }, {});
-      }
-
-      return Hex.stringify(sha1(JSON.stringify(hashParams)));
-    };
-
     const reasons = [];
 
     const gem2sStatus = backendStatus[experimentId]?.status.gem2s?.status;
@@ -118,9 +86,10 @@ const ProjectMenu = () => {
       pipelineStatus.SUCCEEDED, pipelineStatus.RUNNING,
     ].includes(gem2sStatus);
 
-    const paramsHash = backendStatus[experimentId]?.status.gem2s?.paramsHash;
+    const existingParamsHash = backendStatus[experimentId]?.status.gem2s?.paramsHash;
+    const newParamsHash = generateGem2sHashParams(activeProject, samples, experiments[experimentId]);
 
-    const projectHashEqual = paramsHash && paramsHash === generateGem2sHashParams(activeProject);
+    const projectHashEqual = existingParamsHash && existingParamsHash === newParamsHash;
 
     if (!gem2sSuccessful) reasons.push('data has not been processed sucessfully');
     if (!projectHashEqual) reasons.push('the project has been modified');
@@ -133,6 +102,7 @@ const ProjectMenu = () => {
 
   const renderTooltipButton = useCallback(() => {
     const LaunchButton = (props) => {
+      // eslint-disable-next-line react/prop-types
       const { onClick } = props;
 
       return (
@@ -156,7 +126,11 @@ const ProjectMenu = () => {
         <Tooltip
           title={`Ensure all samples are uploaded and all metadata are inserted (no ${DEFAULT_NA})`}
         >
-          <LaunchButton onClick={() => setAnalysisModalVisible(true)} />
+          {/* disabled button inside tooltip causes tooltip to not function */}
+          {/* https://github.com/react-component/tooltip/issues/18#issuecomment-140078802 */}
+          <span>
+            <LaunchButton onClick={() => setAnalysisModalVisible(true)} />
+          </span>
         </Tooltip>
       );
     }
